+++ test/projectile-test.el
@@ -1545,10 +1545,25 @@ You'd normally combine this with `projectile-test-with-sandbox'."
       ;; verify that indexing only invokes these funcs once during recursion
       (spy-on 'projectile-ignored-files :and-call-through)
       (spy-on 'projectile-ignored-directories :and-call-through)
+      (spy-on 'projectile-globally-ignored-directory-names :and-call-through)
 
       (projectile-dir-files-native "projectA/")
       (expect 'projectile-ignored-files :to-have-been-called-times 1)
-      (expect 'projectile-ignored-directories :to-have-been-called-times 1)))))
+      (expect 'projectile-globally-ignored-directory-names :to-have-been-called-times 1)
+      (expect 'projectile-ignored-directories :to-have-been-called-times 1))))
+  (it "ignores globally ignored directories when using native indexing"
+      (projectile-test-with-sandbox
+       (projectile-test-with-files
+        ("project/"
+         "project/.ignoreme/"
+         "project/.ignoreme/should_ignore"
+         "project/src/"
+         "project/src/.ignoreme/"
+         "project/src/.ignoreme/should_ignore"
+         "project/config.conf")
+
+        (setq projectile-globally-ignored-directories '(".ignoreme"))
+        (expect (projectile-dir-files-native "project") :to-equal '("config.conf"))))))
 
 (describe "projectile-process-current-project-buffers-current"
   (it "expects projectile-process-current-project-buffers and