+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Bugs fixed
 
+* [#1250](https://github.com/bbatsov/projectile/issues/1250): `projectile-globally-ignored-directories` not working with native indexing.
 * [#1438](https://github.com/bbatsov/projectile/pull/1438): Make sure `projectile-files-via-ext-command` returns files not errors.
 * [#1450](https://github.com/bbatsov/projectile/pull/1450): Call `switch-project-action` within project's temp buffer.
 * [#1340](https://github.com/bbatsov/projectile/pull/1340): Fix remote projects being removed if TRAMP can't connect.